requirements.txt
snakemake
pandas
fuzzywuzzy
matplotlib
seaborn
pymongo
minio
requests
kafka-python